+++ js/constants.js
@@ -1,0 +1,33 @@
+export const COUNT_MIN_AVATAR = 1;
+export const COUNT_PHOTOS = 25;
+export const MIN_LIKES = 15;
+export const MAX_LIKES = 200;
+export const COUNT_MAX_AVATAR = 6;
+export const COUNT_COMMENTS = 20;
+
+export const DESCRIPTION = [
+  'Красивый пейзаж',
+  'Было здорово! Надо чаще собираться.',
+  'Самый лучший день в моей жизни',
+  'Спасибо всем кто был рядом , это был лучший юбилей',
+  'Я и мои друзья. Люблю их за это.'
+];
+
+export const NAMES = [
+  'Елена',
+  'Вика',
+  'Генадий',
+  'Артем',
+  'Денис',
+  'Таня',
+  'Алена',
+  'Настя',
+  'Игорь',
+  'Константин',
+];
+
+export const MESSAGE = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.'
+];
+
